+++ R/extract_covariates.R
@@ -39,17 +39,19 @@ extract_covariates.random_points <- function(x, covariates, ...) {
 extract_covariates.steps_xy <- function(x, covariates, where = "end", ...) {
   if (class(covariates) %in% paste0("Raster", c("Layer", "Stack", "Brick"))) {
     if (where == "both") {
-      x_start <- raster::extract(covariates, x[, c("x1_", "y1_")], df = TRUE)[, -1, drop = FALSE]
+      x_start <- raster::extract(covariates, as.matrix(x[, c("x1_", "y1_")]),
+                                 df = TRUE)[, -1, drop = FALSE]
       names(x_start) <- paste0(names(x_start), "_start")
-      x_end <- raster::extract(covariates, x[, c("x2_", "y2_")], df = TRUE)[, -1, drop = FALSE]
+      x_end <- raster::extract(covariates, as.matrix(x[, c("x2_", "y2_")]),
+                               df = TRUE)[, -1, drop = FALSE]
       names(x_end) <- paste0(names(x_end), "_end")
       x_all <- cbind(x_start, x_end)
       x[names(x_all)] <- x_all
     } else {
       x[names(covariates)] <- if (where == "end") {
-        raster::extract(covariates, x[, c("x2_", "y2_")])
+        raster::extract(covariates, as.matrix(x[, c("x2_", "y2_")]))
       } else if (where == "start") {
-        raster::extract(covariates, x[, c("x1_", "y1_")])
+        raster::extract(covariates, as.matrix(x[, c("x1_", "y1_")]))
       }
     }
     x