+++ NEWS.md
@@ -1,6 +1,7 @@
 # amt 0.0.5.0
 ## major changes
 - `steps` gains a new argument: `keep_cols` in order to keep columns from point when creating steps.
+- `simulate_tud`: a convenience wrapper arround `simulate_ud` to simulate transition UDs.
 # amt 0.0.4.0
 ## major changes
 - Added an alias to `mk_track` called `make_track`.